refactor: Move KernelStatistics and EndpointStatistics to gql_legacy
