+++ src/ai/backend/manager/models/endpoint/__init__.py
@@ -4,7 +4,6 @@ from ai.backend.manager.data.model_serving.types import EndpointLifecycle
 from .row import (
     EndpointAutoScalingRuleRow,
     EndpointRow,
-    EndpointStatistics,
     EndpointTokenRow,
     ModelServiceHelper,
 )
@@ -15,7 +14,6 @@ __all__ = (
     "EndpointAutoScalingRuleRow",
     "EndpointLifecycle",
     "EndpointRow",
-    "EndpointStatistics",
     "EndpointTokenRow",
     "ModelServiceHelper",
 )
