+++ src/ai/backend/manager/models/endpoint/row.py
@@ -5,7 +5,6 @@ import uuid
 from collections import defaultdict
 from collections.abc import (
     Iterable,
-    Mapping,
     Sequence,
 )
 from datetime import UTC, datetime
@@ -106,16 +105,13 @@ from ai.backend.manager.models.vfolder import prepare_vfolder_mounts
 from ai.backend.manager.types import MountOptionModel, UserScope
 
 if TYPE_CHECKING:
-    from ai.backend.common.clients.valkey_client.valkey_stat.client import ValkeyStatClient
     from ai.backend.manager.data.deployment.creator import DeploymentCreator
     from ai.backend.manager.models.deployment_revision import DeploymentRevisionRow
-    from ai.backend.manager.models.gql import GraphQueryContext
 
 __all__ = (
     "EndpointAutoScalingRuleRow",
     "EndpointLifecycle",
     "EndpointRow",
-    "EndpointStatistics",
     "EndpointTokenRow",
     "ModelServiceHelper",
 )
@@ -1459,33 +1455,3 @@ class ModelServiceHelper:
             ) from e
         except YAMLError as e:
             raise InvalidAPIParameters(f"Invalid YAML syntax: {e}") from e
-
-
-class EndpointStatistics:
-    @classmethod
-    async def batch_load_by_endpoint_impl(
-        cls,
-        valkey_stat_client: ValkeyStatClient,
-        endpoint_ids: Sequence[UUID],
-    ) -> Sequence[Optional[Mapping[str, Any]]]:
-        endpoint_id_strs = [str(endpoint_id) for endpoint_id in endpoint_ids]
-        return await valkey_stat_client.get_inference_app_statistics_batch(endpoint_id_strs)
-
-    @classmethod
-    async def batch_load_by_endpoint(
-        cls,
-        ctx: GraphQueryContext,
-        endpoint_ids: Sequence[UUID],
-    ) -> Sequence[Optional[Mapping[str, Any]]]:
-        return await cls.batch_load_by_endpoint_impl(ctx.valkey_stat, endpoint_ids)
-
-    @classmethod
-    async def batch_load_by_replica(
-        cls,
-        ctx: GraphQueryContext,
-        endpoint_replica_ids: Sequence[tuple[UUID, UUID]],
-    ) -> Sequence[Optional[Mapping[str, Any]]]:
-        endpoint_replica_pairs = [
-            (str(endpoint_id), str(replica_id)) for endpoint_id, replica_id in endpoint_replica_ids
-        ]
-        return await ctx.valkey_stat.get_inference_replica_statistics_batch(endpoint_replica_pairs)
