+++ src/ai/backend/manager/models/kernel/__init__.py
@@ -9,7 +9,6 @@ from .row import (
     RESOURCE_USAGE_KERNEL_STATUSES,
     USER_RESOURCE_OCCUPYING_KERNEL_STATUSES,
     KernelRow,
-    KernelStatistics,
     by_kernel_ids,
     get_user_email,
     handle_kernel_exception,
@@ -26,7 +25,6 @@ __all__ = (
     "RESOURCE_USAGE_KERNEL_STATUSES",
     "USER_RESOURCE_OCCUPYING_KERNEL_STATUSES",
     "KernelRow",
-    "KernelStatistics",
     "KernelStatus",
     "by_kernel_ids",
     "get_user_email",
