+++ src/ai/backend/manager/models/kernel/row.py
@@ -98,9 +98,6 @@ from ai.backend.manager.models.utils import (
     sql_json_merge,
 )
 
-if TYPE_CHECKING:
-    from ai.backend.manager.models.gql import GraphQueryContext
-
 __all__ = (
     "AGENT_RESOURCE_OCCUPYING_KERNEL_STATUSES",
     "DEAD_KERNEL_STATUSES",
@@ -109,7 +106,6 @@ __all__ = (
     "RESOURCE_USAGE_KERNEL_STATUSES",
     "USER_RESOURCE_OCCUPYING_KERNEL_STATUSES",
     "KernelRow",
-    "KernelStatistics",
     "get_user_email",
     "handle_kernel_exception",
     "kernels",
@@ -1084,36 +1080,6 @@ class SessionInfo(TypedDict):
     session_name: str
     status: KernelStatus
     created_at: datetime
-
-
-class KernelStatistics:
-    @classmethod
-    async def batch_load_by_kernel_impl(
-        cls,
-        valkey_stat_client: ValkeyStatClient,
-        session_ids: Sequence[SessionId],
-    ) -> Sequence[Optional[Mapping[str, Any]]]:
-        """For cases where required to collect kernel metrics in bulk internally"""
-        session_ids_str = [str(sess_id) for sess_id in session_ids]
-        return await valkey_stat_client.get_session_statistics_batch(session_ids_str)
-
-    @classmethod
-    async def batch_load_by_kernel(
-        cls,
-        ctx: GraphQueryContext,
-        session_ids: Sequence[SessionId],
-    ) -> Sequence[Optional[Mapping[str, Any]]]:
-        """wrapper of `KernelStatistics.batch_load_by_kernel_impl()` for aiodataloader"""
-        return await cls.batch_load_by_kernel_impl(ctx.valkey_stat, session_ids)
-
-    @classmethod
-    async def batch_load_inference_metrics_by_kernel(
-        cls,
-        ctx: GraphQueryContext,
-        session_ids: Sequence[SessionId],
-    ) -> Sequence[Optional[Mapping[str, Any]]]:
-        session_ids_str = [str(sess_id) for sess_id in session_ids]
-        return await ctx.valkey_live.get_session_statistics_batch(session_ids_str)
 
 
 async def recalc_concurrency_used(
